Refactor Aravis backend: split driver/backend, fix camera capture, hardware-tested
Architecture refactor following STAR pattern:
- cytation_aravis_driver.py: CytationAravisDriver (BioTekBackend + AravisCamera)
  owns connections, optics control, camera control
- cytation_aravis_microscopy.py: CytationAravisMicroscopyBackend (MicroscopyBackend)
  orchestrates capture by calling driver methods
- cytation1.py/cytation5.py: device classes with camera="aravis"|"pyspin" switch
- Removed monolithic cytation_aravis.py

Bugs fixed during hardware testing:
- aravis_camera.py: trigger() no longer starts/stops acquisition per frame —
  matches PySpin pattern where acquisition brackets the capture loop
- cytation_aravis_driver.py: led_on intensity zfill(3) → zfill(2)
- cytation_aravis_driver.py: _load_filters/_load_objectives use Rick's actual
  firmware protocol (per-slot query) instead of made-up regex parsing
- cytation_aravis_driver.py: all serial commands (set_imaging_mode, set_objective,
  set_focus, select, set_position, led_on/off) now match cytation.py exactly
- cytation_aravis_driver.py: bytes→str decode for firmware responses

Hardware-tested on Cytation 1 with BlackFly BFLY-U3-13S2M:
- Single capture: brightfield 2.5x, real images with correct exposure
- Multi-image tiling: 2x2 grid capture working
- Filter/objective discovery: DAPI, GFP, RFP + 4x PL FL, 2.5x FL Zeiss
