+++ pylabrobot/agilent/biotek/__init__.py
@@ -2,5 +2,6 @@ from .biotek import BioTekBackend
 from .cytation import CytationBackend, CytationImagingConfig
 from .cytation1 import Cytation1
 from .cytation5 import Cytation5
+from .cytation_aravis_driver import AravisImagingConfig, CytationAravisDriver
 from .el406 import EL406, EL406Driver, EL406PlateWashingBackend, EL406ShakingBackend
 from .synergy_h1 import SynergyH1, SynergyH1Backend
