+++ pylabrobot/agilent/biotek/cytation5.py
@@ -1,28 +1,25 @@
 """Cytation 5 device — plate reader + imager.
 
-The Cytation 5 combines plate reading (absorbance, fluorescence,
-luminescence) with microscopy imaging. The backend must implement
-both BioTekBackend (serial protocol) and MicroscopyBackend (capture).
-
-Two backend options:
-  - CytationBackend (cytation.py) — uses PySpin for camera
-  - CytationAravisBackend (cytation_aravis.py) — uses Aravis for camera
+Follows the STAR pattern: device creates driver internally based on
+the ``camera`` parameter, setup() wires capabilities.
 
 Example::
 
-    from pylabrobot.agilent.biotek.cytation_aravis import CytationAravisBackend
-    from pylabrobot.agilent.biotek.cytation5 import Cytation5
+    # Aravis (default)
+    cytation = Cytation5(name="cytation5", camera_serial="22580842")
 
-    backend = CytationAravisBackend(camera_serial="22580842")
-    cytation = Cytation5(name="cytation5", backend=backend)
+    # PySpin
+    cytation = Cytation5(name="cytation5", camera="pyspin")
+
     await cytation.setup()
     result = await cytation.microscopy.capture(...)
+    await cytation.stop()
 """
 
 from __future__ import annotations
 
 import logging
-from typing import Optional
+from typing import Literal, Optional
 
 from pylabrobot.capabilities.microscopy import Microscopy
 from pylabrobot.capabilities.plate_reading.absorbance import Absorbance
@@ -32,32 +29,48 @@ from pylabrobot.capabilities.temperature_controlling import TemperatureControlle
 from pylabrobot.device import Device
 from pylabrobot.resources import Coordinate, PlateHolder, Resource
 
-from .biotek import BioTekBackend
-
 logger = logging.getLogger(__name__)
 
 
 class Cytation5(Resource, Device):
   """Agilent BioTek Cytation 5 — plate reader + imager.
 
-  Takes a backend that provides both plate reading (via BioTekBackend)
-  and microscopy (via MicroscopyBackend). Pass either CytationBackend
-  (PySpin) or CytationAravisBackend (Aravis).
+  Creates the appropriate driver based on the ``camera`` parameter:
+    - ``"aravis"`` (default): CytationAravisDriver (Aravis/GenICam)
+    - ``"pyspin"``: CytationBackend (PySpin/Spinnaker SDK)
 
   Capabilities:
     - absorbance, fluorescence, luminescence (plate reading)
-    - microscopy (imaging via camera)
+    - microscopy (imaging)
     - temperature (incubation)
   """
 
   def __init__(
     self,
     name: str,
-    backend: BioTekBackend,
+    camera: Literal["aravis", "pyspin"] = "aravis",
+    camera_serial: Optional[str] = None,
+    device_id: Optional[str] = None,
     size_x: float = 0.0,
     size_y: float = 0.0,
     size_z: float = 0.0,
   ):
+    if camera == "aravis":
+      from .cytation_aravis_driver import CytationAravisDriver
+      driver = CytationAravisDriver(
+        camera_serial=camera_serial,
+        device_id=device_id,
+      )
+    elif camera == "pyspin":
+      from .cytation import CytationBackend, CytationImagingConfig
+      config = CytationImagingConfig(camera_serial_number=camera_serial)
+      driver = CytationBackend(
+        device_id=device_id,
+        imaging_config=config,
+      )
+    else:
+      raise ValueError(f"Unknown camera backend: {camera!r}. Use 'aravis' or 'pyspin'.")
+
     Resource.__init__(
       self,
       name=name,
@@ -66,21 +79,15 @@ class Cytation5(Resource, Device):
       size_z=size_z,
       model="Agilent BioTek Cytation 5",
     )
-    Device.__init__(self, driver=backend)
-    self.driver: BioTekBackend = backend
+    Device.__init__(self, driver=driver)
+    self.driver = driver
+    self._camera = camera
 
-    self.absorbance = Absorbance(backend=backend)
-    self.luminescence = Luminescence(backend=backend)
-    self.fluorescence = Fluorescence(backend=backend)
-    self.microscopy = Microscopy(backend=backend)
-    self.temperature = TemperatureController(backend=backend)
-    self._capabilities = [
-      self.absorbance,
-      self.luminescence,
-      self.fluorescence,
-      self.microscopy,
-      self.temperature,
-    ]
+    self.absorbance: Absorbance  # set in setup()
+    self.luminescence: Luminescence  # set in setup()
+    self.fluorescence: Fluorescence  # set in setup()
+    self.microscopy: Microscopy  # set in setup()
+    self.temperature: TemperatureController  # set in setup()
 
     self.plate_holder = PlateHolder(
       name=name + "_plate_holder",
@@ -91,6 +98,41 @@ class Cytation5(Resource, Device):
       child_location=Coordinate.zero(),
     )
     self.assign_child_resource(self.plate_holder, location=Coordinate.zero())
+
+  async def setup(self) -> None:
+    if self._camera == "aravis":
+      await self.driver.setup()
+      self.microscopy = Microscopy(backend=self.driver.microscopy_backend)
+    else:
+      await self.driver.setup(use_cam=True)
+      self.microscopy = Microscopy(backend=self.driver)
+
+    # Plate reading + temperature use the driver directly
+    # (BioTekBackend implements these backend ABCs)
+    self.absorbance = Absorbance(backend=self.driver)
+    self.luminescence = Luminescence(backend=self.driver)
+    self.fluorescence = Fluorescence(backend=self.driver)
+    self.temperature = TemperatureController(backend=self.driver)
+
+    self._capabilities = [
+      self.absorbance,
+      self.luminescence,
+      self.fluorescence,
+      self.microscopy,
+      self.temperature,
+    ]
+
+    for cap in self._capabilities:
+      await cap._on_setup()
+    self._setup_finished = True
+    logger.info("Cytation5 setup complete (camera=%s)", self._camera)
+
+  async def stop(self) -> None:
+    for cap in reversed(self._capabilities):
+      await cap._on_stop()
+    await self.driver.stop()
+    self._setup_finished = False
+    logger.info("Cytation5 stopped")
 
   def serialize(self) -> dict:
     return {**Resource.serialize(self), **Device.serialize(self)}
